GeminiProject/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.27
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-dotenv==1.0.0
google-generativeai==0.3.2
pytest-mock==3.12.0